fix : 11058 - 크리보드 dp 배열 타입 변경

+++ src/main/java/menz/study/week07/JeongSeok/_11058_.java
@@ -11,7 +11,7 @@ public class _11058_ {
 
 		int N = Integer.parseInt(br.readLine());
 		
-		int[] dp = new int[N + 1];
+		long[] dp = new long[N + 1];
 
 		for (int i = 1; i <= 6; i++) {
 			dp[i] = i;
